mali-t6xx: fix null pointer deref whn dumping MMU page tables
At the moment when if we perform an instrumented dump of the MMU page tables,
the munmap call back will try to free physical pages that are supposed to be
in the memory region allocated for the dump but that are not there,
thus resulting in a NULL pointer exception followed by kernel panic.

In order to solve this, in "kbase_free_phy_pages()" (that is called by the
munmap callback) we need to check the memory region flags to see if
KBASE_REG_IS_MMU_DUMP flag is set.  If it is set then we have to skip the
 call to free the physical pages.

Bug:  7432454
Signed-off-by: Gabriele Paoloni <gabriele.paoloni@arm.com>
Change-Id: I37d5c9f634ae08a1e8340671b15287354c75fd64
Signed-off-by: franciscofranco <franciscofranco.1990@gmail.com>